refactor: how children elements are display in Header component
This option is to display extra page options in the form of a
menu. We now treat the childre as possible array of Nodes
using the recommended React.Children.map method to loop
over React children nodes. All items display and a list
of items to create a menu like effect.